Add files via upload
CC6. Loops. Build a dashboard. 
Use the ONS API to batch download nine different series as JSON files
using a loop. 
Save these to your GitHub account, and use these (as “raw” files) to 
supply the data to nine (or more) charts on a theme or themes of your 
choice.